+++ app/src/main/java/com/example/ethan/share01/RbPreference.java
@@ -21,8 +21,8 @@ public class RbPreference {
     }
 
     public void put(String key, String value) {
-        SharedPreferences pref = mContext.getSharedPreferences(PREF_NAME,
-                Activity.MODE_PRIVATE);
+        SharedPreferences pref = mContext.getApplicationContext().getSharedPreferences(PREF_NAME,
+                Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = pref.edit();
 
         editor.putString(key, value);
@@ -30,8 +30,8 @@ public class RbPreference {
     }
 
     public void put(String key, boolean value) {
-        SharedPreferences pref = mContext.getSharedPreferences(PREF_NAME,
-                Activity.MODE_PRIVATE);
+        SharedPreferences pref = mContext.getApplicationContext().getSharedPreferences(PREF_NAME,
+                Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = pref.edit();
 
         editor.putBoolean(key, value);
@@ -39,8 +39,8 @@ public class RbPreference {
     }
 
     public void put(String key, int value) {
-        SharedPreferences pref = mContext.getSharedPreferences(PREF_NAME,
-                Activity.MODE_PRIVATE);
+        SharedPreferences pref = mContext.getApplicationContext().getSharedPreferences(PREF_NAME,
+                Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = pref.edit();
 
         editor.putInt(key, value);
@@ -48,8 +48,8 @@ public class RbPreference {
     }
 
     public String getValue(String key, String dftValue) {
-        SharedPreferences pref = mContext.getSharedPreferences(PREF_NAME,
-                Activity.MODE_PRIVATE);
+        SharedPreferences pref = mContext.getApplicationContext().getSharedPreferences(PREF_NAME,
+                Context.MODE_PRIVATE);
 
         try {
             return pref.getString(key, dftValue);
@@ -60,8 +60,8 @@ public class RbPreference {
     }
 
     public int getValue(String key, int dftValue) {
-        SharedPreferences pref = mContext.getSharedPreferences(PREF_NAME,
-                Activity.MODE_PRIVATE);
+        SharedPreferences pref = mContext.getApplicationContext().getSharedPreferences(PREF_NAME,
+                Context.MODE_PRIVATE);
 
         try {
             return pref.getInt(key, dftValue);
@@ -72,8 +72,8 @@ public class RbPreference {
     }
 
     public boolean getValue(String key, boolean dftValue) {
-        SharedPreferences pref = mContext.getSharedPreferences(PREF_NAME,
-                Activity.MODE_PRIVATE);
+        SharedPreferences pref = mContext.getApplicationContext().getSharedPreferences(PREF_NAME,
+                Context.MODE_PRIVATE);
 
         try {
             return pref.getBoolean(key, dftValue);
@@ -83,8 +83,8 @@ public class RbPreference {
     }
 
     public void delValue(String key) {
-        SharedPreferences pref = mContext.getSharedPreferences(PREF_NAME,
-                Activity.MODE_PRIVATE);
+        SharedPreferences pref = mContext.getApplicationContext().getSharedPreferences(PREF_NAME,
+                Context.MODE_PRIVATE);
 
         SharedPreferences.Editor editor = pref.edit();
 
@@ -94,8 +94,8 @@ public class RbPreference {
     }
 
     public void removeAllValue() {
-        SharedPreferences pref = mContext.getSharedPreferences(PREF_NAME,
-                Activity.MODE_PRIVATE);
+        SharedPreferences pref = mContext.getApplicationContext().getSharedPreferences(PREF_NAME,
+                Context.MODE_PRIVATE);
 
         SharedPreferences.Editor editor = pref.edit();
         editor.clear();
